HTBHF-2011 Split report properties into two separate classes, one for claims, one for payments.
