+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/reporting/payload/ReportPropertiesFactory.java
@@ -1,14 +1,7 @@
 package uk.gov.dhsc.htbhf.claimant.reporting.payload;
 
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
-import uk.gov.dhsc.htbhf.claimant.entity.PaymentCycle;
-import uk.gov.dhsc.htbhf.claimant.message.context.ReportClaimMessageContext;
 import uk.gov.dhsc.htbhf.claimant.message.context.ReportEventMessageContext;
-import uk.gov.dhsc.htbhf.claimant.message.context.ReportPaymentMessageContext;
-import uk.gov.dhsc.htbhf.claimant.message.processor.ChildDateOfBirthCalculator;
-import uk.gov.dhsc.htbhf.claimant.message.processor.NextPaymentCycleSummary;
 import uk.gov.dhsc.htbhf.claimant.model.PostcodeData;
 import uk.gov.dhsc.htbhf.claimant.reporting.ClaimantCategoryCalculator;
 
@@ -16,16 +9,17 @@ import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.Period;
-import java.time.temporal.ChronoUnit;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.TreeMap;
 
-import static uk.gov.dhsc.htbhf.claimant.message.processor.ChildDateOfBirthCalculator.*;
+import static uk.gov.dhsc.htbhf.claimant.message.processor.ChildDateOfBirthCalculator.getNumberOfChildrenUnderFour;
+import static uk.gov.dhsc.htbhf.claimant.message.processor.ChildDateOfBirthCalculator.getNumberOfChildrenUnderOne;
 import static uk.gov.dhsc.htbhf.claimant.reporting.payload.CustomDimension.*;
-import static uk.gov.dhsc.htbhf.claimant.reporting.payload.CustomMetric.*;
-import static uk.gov.dhsc.htbhf.claimant.reporting.payload.EventCategory.CLAIM;
-import static uk.gov.dhsc.htbhf.claimant.reporting.payload.EventCategory.PAYMENT;
+import static uk.gov.dhsc.htbhf.claimant.reporting.payload.CustomMetric.CHILDREN_BETWEEN_ONE_AND_FOUR;
+import static uk.gov.dhsc.htbhf.claimant.reporting.payload.CustomMetric.CHILDREN_UNDER_ONE;
+import static uk.gov.dhsc.htbhf.claimant.reporting.payload.CustomMetric.CLAIMANT_AGE;
+import static uk.gov.dhsc.htbhf.claimant.reporting.payload.CustomMetric.PREGNANCIES;
 import static uk.gov.dhsc.htbhf.claimant.reporting.payload.EventProperties.*;
 import static uk.gov.dhsc.htbhf.claimant.reporting.payload.MandatoryProperties.HIT_TYPE_KEY;
 import static uk.gov.dhsc.htbhf.claimant.reporting.payload.MandatoryProperties.PROTOCOL_VERSION_KEY;
@@ -35,9 +29,7 @@ import static uk.gov.dhsc.htbhf.claimant.reporting.payload.UserType.ONLINE;
 /**
  * Factory class for creating a map of parameters reported to google analytics measurement protocol.
  */
-@Component
-@SuppressWarnings("PMD.TooManyMethods")
-public class ReportPropertiesFactory {
+public abstract class ReportPropertiesFactory {
 
     private static final String HIT_TYPE_VALUE = "event";
     private static final String PROTOCOL_VERSION_VALUE = "1";
@@ -47,37 +39,26 @@ public class ReportPropertiesFactory {
     public static final int PREGNANCY_DURATION_IN_WEEKS = 40;
 
     private final ClaimantCategoryCalculator claimantCategoryCalculator;
-    private final ChildDateOfBirthCalculator childDateOfBirthCalculator;
     private final String trackingId;
 
-    public ReportPropertiesFactory(@Value("${google-analytics.tracking-id}") String trackingId,
-                                   ClaimantCategoryCalculator claimantCategoryCalculator,
-                                   ChildDateOfBirthCalculator childDateOfBirthCalculator) {
+    public ReportPropertiesFactory(String trackingId,
+                                   ClaimantCategoryCalculator claimantCategoryCalculator) {
         this.trackingId = trackingId;
         this.claimantCategoryCalculator = claimantCategoryCalculator;
-        this.childDateOfBirthCalculator = childDateOfBirthCalculator;
     }
 
-    public Map<String, String> createReportPropertiesForClaimEvent(ReportClaimMessageContext context) {
+    protected Long calculateQueueTime(LocalDateTime timestamp) {
+        long queueTime = Duration.between(timestamp, LocalDateTime.now()).toMillis();
+        return Math.min(queueTime, MAX_QUEUE_TIME);
+    }
+
+    protected Map<String, String> mapValuesToString(Map<String, Object> original) {
         Map<String, String> reportProperties = new LinkedHashMap<>();
-        reportProperties.putAll(mapValuesToString(createMandatoryPropertiesMap()));
-        reportProperties.putAll(mapValuesToString(createEventPropertiesMap(context, CLAIM, 0)));
-        reportProperties.putAll(mapValuesToString(createCustomDimensionMap(context)));
-        reportProperties.putAll(mapValuesToString(createCustomMetricMapForClaimEvent(context)));
+        original.forEach((key, value) -> reportProperties.put(key, String.valueOf(value)));
         return reportProperties;
     }
 
-    public Map<String, String> createReportPropertiesForPaymentEvent(ReportPaymentMessageContext context) {
-        Map<String, String> reportProperties = new LinkedHashMap<>();
-        reportProperties.putAll(mapValuesToString(createMandatoryPropertiesMap()));
-        int totalPaymentAmount = context.getPaymentForPregnancy() + context.getPaymentForChildrenUnderOne() + context.getPaymentForChildrenBetweenOneAndFour();
-        reportProperties.putAll(mapValuesToString(createEventPropertiesMap(context, PAYMENT, totalPaymentAmount)));
-        reportProperties.putAll(mapValuesToString(createCustomDimensionMap(context)));
-        reportProperties.putAll(mapValuesToString(createCustomMetricMapForPaymentEvent(context)));
-        return reportProperties;
-    }
-
-    private Map<String, Object> createMandatoryPropertiesMap() {
+    protected Map<String, Object> createMandatoryPropertiesMap() {
         Map<String, Object> mandatoryProperties = new LinkedHashMap<>();
         mandatoryProperties.put(HIT_TYPE_KEY.getFieldName(), HIT_TYPE_VALUE);
         mandatoryProperties.put(PROTOCOL_VERSION_KEY.getFieldName(), PROTOCOL_VERSION_VALUE);
@@ -85,7 +66,7 @@ public class ReportPropertiesFactory {
         return mandatoryProperties;
     }
 
-    private Map<String, Object> createEventPropertiesMap(ReportEventMessageContext context, EventCategory eventCategory, int eventValue) {
+    protected Map<String, Object> createEventPropertiesMap(ReportEventMessageContext context, EventCategory eventCategory, int eventValue) {
         Map<String, Object> eventPropertiesMap = new LinkedHashMap<>();
         eventPropertiesMap.put(EVENT_CATEGORY.getFieldName(), eventCategory.name());
         eventPropertiesMap.put(EVENT_ACTION.getFieldName(), context.getEventAction());
@@ -95,7 +76,7 @@ public class ReportPropertiesFactory {
         return eventPropertiesMap;
     }
 
-    private Map<String, Object> createCustomDimensionMap(ReportEventMessageContext context) {
+    protected Map<String, Object> createCustomDimensionMap(ReportEventMessageContext context) {
         Map<String, Object> customDimensions = new TreeMap<>();
         customDimensions.put(USER_TYPE.getFieldName(), ONLINE.name());
         ClaimantCategory claimantCategory = claimantCategoryCalculator
@@ -112,24 +93,7 @@ public class ReportPropertiesFactory {
         return customDimensions;
     }
 
-    private Map<String, Object> createCustomMetricMapForClaimEvent(ReportClaimMessageContext context) {
-        Map<String, Object> customMetrics = createCommonCustomMetrics(context);
-        LocalDate expectedDeliveryDate = context.getClaim().getClaimant().getExpectedDeliveryDate();
-        LocalDate atDate = context.getTimestamp().toLocalDate();
-        if (isClaimantPregnant(expectedDeliveryDate, atDate)) {
-            LocalDate conception = expectedDeliveryDate.minusWeeks(PREGNANCY_DURATION_IN_WEEKS);
-            customMetrics.put(WEEKS_PREGNANT.getFieldName(), ChronoUnit.WEEKS.between(conception, atDate));
-        }
-        return customMetrics;
-    }
-
-    private Map<String, Object> createCustomMetricMapForPaymentEvent(ReportPaymentMessageContext context) {
-        Map<String, Object> customMetrics = createCommonCustomMetrics(context);
-        addPaymentCycleMetrics(context, customMetrics);
-        return customMetrics;
-    }
-
-    private Map<String, Object> createCommonCustomMetrics(ReportEventMessageContext context) {
+    protected Map<String, Object> createCommonCustomMetrics(ReportEventMessageContext context) {
         Map<String, Object> customMetrics = new TreeMap<>();
         Claimant claimant = context.getClaim().getClaimant();
         LocalDate atDate = context.getTimestamp().toLocalDate();
@@ -144,27 +108,7 @@ public class ReportPropertiesFactory {
         return customMetrics;
     }
 
-    private boolean isClaimantPregnant(LocalDate expectedDeliveryDate, LocalDate atDate) {
+    protected boolean isClaimantPregnant(LocalDate expectedDeliveryDate, LocalDate atDate) {
         return expectedDeliveryDate != null && !expectedDeliveryDate.isBefore(atDate);
-    }
-
-    private void addPaymentCycleMetrics(ReportPaymentMessageContext context, Map<String, Object> customMetrics) {
-        PaymentCycle paymentCycle = context.getPaymentCycle();
-        customMetrics.put(PAYMENT_FOR_PREGNANCY.getFieldName(), context.getPaymentForPregnancy());
-        customMetrics.put(PAYMENT_FOR_CHILDREN_UNDER_ONE.getFieldName(), context.getPaymentForChildrenUnderOne());
-        customMetrics.put(PAYMENT_FOR_CHILDREN_BETWEEN_ONE_AND_FOUR.getFieldName(), context.getPaymentForChildrenBetweenOneAndFour());
-        NextPaymentCycleSummary nextPaymentCycleSummary = childDateOfBirthCalculator.getNextPaymentCycleSummary(paymentCycle);
-        customMetrics.put(NUMBER_OF_1ST_OR_FOURTH_BIRTHDAYS_IN_NEXT_CYCLE.getFieldName(), nextPaymentCycleSummary.getNumberOfChildrenTurningOneOrFour());
-    }
-
-    private Long calculateQueueTime(LocalDateTime timestamp) {
-        long queueTime = Duration.between(timestamp, LocalDateTime.now()).toMillis();
-        return Math.min(queueTime, MAX_QUEUE_TIME);
-    }
-
-    private Map<String, String> mapValuesToString(Map<String, Object> original) {
-        Map<String, String> reportProperties = new LinkedHashMap<>();
-        original.forEach((key, value) -> reportProperties.put(key, String.valueOf(value)));
-        return reportProperties;
     }
 }
