+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/reporting/MIReporterTest.java
@@ -9,7 +9,8 @@ import uk.gov.dhsc.htbhf.claimant.entity.Claim;
 import uk.gov.dhsc.htbhf.claimant.message.context.ReportClaimMessageContext;
 import uk.gov.dhsc.htbhf.claimant.message.context.ReportPaymentMessageContext;
 import uk.gov.dhsc.htbhf.claimant.model.PostcodeData;
-import uk.gov.dhsc.htbhf.claimant.reporting.payload.ReportPropertiesFactory;
+import uk.gov.dhsc.htbhf.claimant.reporting.payload.ReportClaimPropertiesFactory;
+import uk.gov.dhsc.htbhf.claimant.reporting.payload.ReportPaymentPropertiesFactory;
 import uk.gov.dhsc.htbhf.claimant.repository.ClaimRepository;
 
 import java.util.Map;
@@ -40,7 +41,9 @@ class MIReporterTest {
     @Mock
     private ClaimRepository claimRepository;
     @Mock
-    private ReportPropertiesFactory reportPropertiesFactory;
+    private ReportClaimPropertiesFactory reportClaimPropertiesFactory;
+    @Mock
+    private ReportPaymentPropertiesFactory reportPaymentPropertiesFactory;
     @Mock
     private GoogleAnalyticsClient googleAnalyticsClient;
 
@@ -54,14 +57,14 @@ class MIReporterTest {
         String postcode = claim.getClaimant().getAddress().getPostcode();
         PostcodeData postcodeData = aPostcodeDataObjectForPostcode(postcode);
         given(postcodeDataClient.getPostcodeData(any())).willReturn(postcodeData);
-        given(reportPropertiesFactory.createReportPropertiesForClaimEvent(any())).willReturn(REPORT_PROPERTIES);
+        given(reportClaimPropertiesFactory.createReportPropertiesForClaimEvent(any())).willReturn(REPORT_PROPERTIES);
 
         miReporter.reportClaim(context);
 
         assertThat(claim.getPostcodeData()).isEqualTo(postcodeData);
         verify(postcodeDataClient).getPostcodeData(claim);
         verify(claimRepository).save(claim);
-        verify(reportPropertiesFactory).createReportPropertiesForClaimEvent(context);
+        verify(reportClaimPropertiesFactory).createReportPropertiesForClaimEvent(context);
         verify(googleAnalyticsClient).reportEvent(REPORT_PROPERTIES);
     }
 
@@ -70,12 +73,12 @@ class MIReporterTest {
         PostcodeData postcodeData = aPostcodeDataObjectForPostcode(VALID_POSTCODE);
         Claim claim = aClaimWithPostcodeData(postcodeData);
         ReportClaimMessageContext context = ReportClaimMessageContext.builder().claim(claim).build();
-        given(reportPropertiesFactory.createReportPropertiesForClaimEvent(any())).willReturn(REPORT_PROPERTIES);
+        given(reportClaimPropertiesFactory.createReportPropertiesForClaimEvent(any())).willReturn(REPORT_PROPERTIES);
 
         miReporter.reportClaim(context);
 
         verifyZeroInteractions(postcodeDataClient, claimRepository);
-        verify(reportPropertiesFactory).createReportPropertiesForClaimEvent(context);
+        verify(reportClaimPropertiesFactory).createReportPropertiesForClaimEvent(context);
         verify(googleAnalyticsClient).reportEvent(REPORT_PROPERTIES);
     }
 
@@ -85,13 +88,13 @@ class MIReporterTest {
         PostcodeData postcodeData = aPostcodeDataObjectForPostcode(VALID_POSTCODE);
         ReportPaymentMessageContext context = aReportPaymentMessageContextWithClaim(claim);
         given(postcodeDataClient.getPostcodeData(any())).willReturn(postcodeData);
-        given(reportPropertiesFactory.createReportPropertiesForPaymentEvent(any())).willReturn(REPORT_PROPERTIES);
+        given(reportPaymentPropertiesFactory.createReportPropertiesForPaymentEvent(any())).willReturn(REPORT_PROPERTIES);
 
         miReporter.reportPayment(context);
 
         verify(postcodeDataClient).getPostcodeData(claim);
         verify(claimRepository).save(claim);
-        verify(reportPropertiesFactory).createReportPropertiesForPaymentEvent(context);
+        verify(reportPaymentPropertiesFactory).createReportPropertiesForPaymentEvent(context);
         verify(googleAnalyticsClient).reportEvent(REPORT_PROPERTIES);
     }
 
@@ -100,12 +103,12 @@ class MIReporterTest {
         PostcodeData postcodeData = aPostcodeDataObjectForPostcode(VALID_POSTCODE);
         Claim claim = aClaimWithPostcodeData(postcodeData);
         ReportPaymentMessageContext context = aReportPaymentMessageContextWithClaim(claim);
-        given(reportPropertiesFactory.createReportPropertiesForPaymentEvent(any())).willReturn(REPORT_PROPERTIES);
+        given(reportPaymentPropertiesFactory.createReportPropertiesForPaymentEvent(any())).willReturn(REPORT_PROPERTIES);
 
         miReporter.reportPayment(context);
 
         verifyZeroInteractions(postcodeDataClient, claimRepository);
-        verify(reportPropertiesFactory).createReportPropertiesForPaymentEvent(context);
+        verify(reportPaymentPropertiesFactory).createReportPropertiesForPaymentEvent(context);
         verify(googleAnalyticsClient).reportEvent(REPORT_PROPERTIES);
     }
 }
